--- src/api/validations.js
@@ -1,15 +1,0 @@
-export const RegisterValidate = {
-    errors: [
-        {400: "Preencha os dados obrigatórios"},
-        {401: "Usuário não autenticado"},
-        {403: "Email já cadastrado"},
-    ]
-};
-
-export const LoginValidate = {
-    errors: [
-        {400: "E-mail ou senha inválidos"},
-        {401: "Usuário não autenticado"},
-        {404: "Usuário não encontrado"},
-    ]
-};
